Fully update to C23; *GCC 15+ or Clang 20+ is now required!*
It was already std=c2x so i guess not much changed, but this is just to 
guarantee support for certain features.

In theory GCC 14+ should be okay for the 32-bit ELF and x86-64 PE 
toolchains, since I don't plan to use #embed there

But I'll add more info in the readme later on.