Fix MBR issues (still incomplete), add MBR / make partition, use sfdisk
Still need to implement my own MBR (as well as deal with.. well, 
everything else, really. this whole process took me 5 hours and it's 
basically duct tape. how is it so hard??)

Not to mention the complete and utter lack of GPT support (though at 
least you can just use the regular Partition.img on EFI, so that's fine)

I'm just annoyed by how stupid and convoluted this whole process is.. 
this might be a step in the right direction but it involved following so 
many wrong steps that I'm starting to regret it

At least, there theoretically isn't any issue with the boot code, the 
ridiculous hassle is writing a partition table without cobbling together 
a bunch of stupid workarounds

At least it works ugghfdjgfdgfdg