Fix code problems (I don't know if this fixes any bugs but i hope so)
Passes cppcheck (mostly) and GCC's -fanalyzer now.

I have a PC (that I can't test on right now) that was crashing right as I was switching modes; I knew it could switch to the kernel fine, so I think it had to do with this code

But I don't know, I hope that fixed it though.