Fix bugs, make the BIOS loader properly load EFI apps, update SSE funcs
In theory, both BIOS and EFI loaders should now load the kernel executable without any issue

Previously, I was just reading the .text section directly off of the file.. which can sometimes work, but the moment you need to access something *outside* of the .text section and you don't have access to it, it becomes an issue.

This fixes that, as well as a few other bugs I hadn't noticed. Hopefully this works well, because this is likely the last thing I'll add to *either* of the loaders