Check for etDyn (not etExec) when reading the kernel ELF headers
Long story short, *actual* position-independent executables are linked 
as ET_DYN (a shared object), not ET_EXEC - I was previously just 
checking for the latter, when it should be the former.

Additionally, there's a few extra things I need to do - namely, I need 
to load dynamic program headers as well (which I wasn't doing before..), 
as well as properly page-align it in the linker.

(To be perfectly honest I don't know if this is completely correct, but 
it seems to work fine - this is a little hard to wrap my head around, I 
just hope this is okay)