Fix a really stupid bug
Long story short - I had a regular 32-bit GDT descriptor, with a 4-byte 
address instead of an 8-byte one.

Usually that wouldn't be an issue - and it's guaranteed not to be if 
you're using protected mode - but if you're switching from 64-bit (long) 
to 32-bit (compatibility) mode, the `lgdt` instruction expects a 64-bit 
address (`dq` instead of `dd` in NASM).

I previously used dd in Boot/Legacy/Stage3/Stub.asm, and it was fine, 
because the 4 bytes after were the null segment of our GDT, so they were 
essentially zero; but that's not the case in Int13.asm

So, it was indirectly reading the 4-byte address, and *then* the next 
four bytes, which in this case happened to be the GDT descriptor 
afterwards.

That resulted in an error where it was complaining about the GDT address 
not being canonical, because it had something like 020154710007FE70h 
instead of 000000000007FE70h

I love x86 man.