Add Memmove() and fix ReadDisk() - fully functional now, thank god!!
Long story short - my code was mostly fine, the issue is that I was using Memcpy() since I hadn't implemented Memmove() yet.

I was under the illusion that my Memcpy code worked just as well for overlapping memory regions (which is what Memmove is), and it.. *almost* does, but with the tiny source/destination offsets I was using, it didn't

It never crashed though, so I just assumed it was an offset issue of some sort, which I'm pretty sure it isn't.

Now though, it works pretty well; I have a demo up for each device that shows red if there's ever a mismatch between what reading the sectors 'manually' and ReadDisk() shows, and so far, it's always been fine