Add Memcmp() to kernel, and fix all bool-returning Memcmps to return int
I was setting a really dangerous precedent earlier on, where memcmp 
would return bool - which.. is an issue, because it converts cleanly to 
int, but there it's the opposite!

I also had to modify a few checks in the EFI bootloader code to account 
for the updated Memcmp (the even worse thing is that it did the same 
thing there, without even so much as warning! man the C standard is so 
bad)

But it should be fine now. (plus, the kernel having memcmp means it's 
now prepared for any calls the compiler might make, something I was 
really missing)