Properly preserve 64-bit registers and rbp when switching to real mode
I assumed that the extended registers (r12 through r15) were preserved when you switched in and out of long mode, but it seems like that isn't the case on some PCs

It still works fine on most computers, but it didn't on my main desktop (a Ryzen PC). Maybe it trashes the extended 64-bit registers or uses them in int 13h?